# server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Airbnb Visualisation API",
    "description": "API for accessing and analyzing Airbnb listing data",
    "version": "1.0.0",
    "contact": {
      "name": "Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/listings": {
      "get": {
        "summary": "Get paginated listings",
        "description": "Returns a paginated list of Airbnb listings with basic information",
        "tags": [
          "Listings"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number for pagination (default is 1)"
          },
          {
            "in": "query",
            "name": "page_size",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "description": "Number of items per page (default is 10, max 100)"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of Airbnb listings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "Unique identifier of the listing"
                      },
                      "name": {
                        "type": "string",
                        "description": "Name of the listing"
                      },
                      "neighbourhood": {
                        "type": "string",
                        "description": "Neighbourhood where the listing is located"
                      },
                      "room_type": {
                        "type": "string",
                        "description": "Type of room (e.g., Entire home/apt, Private room)"
                      },
                      "price": {
                        "type": "number",
                        "format": "float",
                        "description": "Price per night"
                      },
                      "number_of_reviews": {
                        "type": "integer",
                        "description": "Number of reviews for the listing"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid pagination parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Database error. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/listings/{listing_id}": {
      "get": {
        "summary": "Get a specific listing by ID",
        "description": "Returns detailed information about a specific Airbnb listing",
        "tags": [
          "Listings"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Unique identifier of the listing"
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed information about a listing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "host_id": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "picture_url": {
                      "type": "string"
                    },
                    "neighbourhood_cleansed": {
                      "type": "string"
                    },
                    "latitude": {
                      "type": "number",
                      "format": "float"
                    },
                    "longitude": {
                      "type": "number",
                      "format": "float"
                    },
                    "accommodates": {
                      "type": "integer"
                    },
                    "bathrooms": {
                      "type": "number",
                      "format": "float"
                    },
                    "bedrooms": {
                      "type": "number",
                      "format": "float"
                    },
                    "beds": {
                      "type": "number",
                      "format": "float"
                    },
                    "price": {
                      "type": "number",
                      "format": "float"
                    },
                    "room_type_simple": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid listing ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Listing not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Listing not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Database error. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/listings/{listing_id}/reviews": {
      "get": {
        "summary": "Get reviews for a specific listing",
        "description": "Returns paginated reviews for a specific Airbnb listing",
        "tags": [
          "Listings",
          "Reviews"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Unique identifier of the listing"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number for pagination (default is 1)"
          },
          {
            "in": "query",
            "name": "page_size",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "description": "Number of items per page (default is 10, max 100)"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of reviews for the specified listing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "description": "Unique identifier of the review"
                      },
                      "listing_id": {
                        "type": "integer",
                        "description": "ID of the listing this review belongs to"
                      },
                      "date": {
                        "type": "string",
                        "format": "date",
                        "description": "Date the review was posted"
                      },
                      "reviewer_id": {
                        "type": "integer",
                        "description": "ID of the reviewer"
                      },
                      "reviewer_name": {
                        "type": "string",
                        "description": "Name of the reviewer"
                      },
                      "comments": {
                        "type": "string",
                        "description": "Review text"
                      },
                      "sentiment": {
                        "type": "string",
                        "enum": [
                          "Positive",
                          "Neutral",
                          "Negative"
                        ],
                        "description": "Sentiment analysis of the review"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Database error. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}